dressing up :feelsgood:

+++ core/src/main/java/me/arcademadness/omni_dungeon/attributes/ActionPoints.java
@@ -1,7 +1,7 @@
 package me.arcademadness.omni_dungeon.attributes;
 
-public final class ActionPoints extends SimpleAttribute<Double> {
-    public ActionPoints(double baseValue) {
+public final class ActionPoints extends SimpleAttribute<Integer> {
+    public ActionPoints(int baseValue) {
         super(baseValue);
     }
 }
